Reuse the physical test fixture and fetch the NCSX bundle explicitly
